Support to inference on LiveLineHeatmapViewController and remove LineImageViewController
